Feature added (Scanner): Added notion of a PATH scanner
* FW_PATH_SCANNER: A class that knows how to recursively scan file
system path. Deferred, so the scanner must be implemented in the client
for the particular "things" is it scanning for.
* ECF: Changed description
* TEST_SETs: Various tests added as well as support files. There is a
false positive test for the SET PATH due to some yet to be known
effective code that will allow us to SET and check the PATH string from
the OS (Windows now--Linux/Mac later?)